% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TTformat.r
\name{formatPercentage}
\alias{formatPercentage}
\title{Format table columns}
\usage{
formatPercentage(
  table,
  columns,
  digits = 0,
  interval = 3,
  mark = ",",
  dec.mark = getOption("OutDec")
)
}
\description{
\code{\link[DT]{formatPercentage}}
}
