% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TTformat.r
\name{formatRound}
\alias{formatRound}
\title{Format table columns}
\usage{
formatRound(
  table,
  columns,
  digits = 2,
  interval = 3,
  mark = ",",
  dec.mark = getOption("OutDec")
)
}
\description{
\code{\link[DT]{formatRound}}
}
