% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TTformat.r
\name{formatSignif}
\alias{formatSignif}
\title{Format table columns}
\usage{
formatSignif(
  table,
  columns,
  digits = 2,
  interval = 3,
  mark = ",",
  dec.mark = getOption("OutDec")
)
}
\description{
\code{\link[DT]{formatSignif}}
}
