% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TTformat.r
\name{formatStyle}
\alias{formatStyle}
\title{Format table columns}
\usage{
formatStyle(
  table,
  columns,
  valueColumns = columns,
  target = c("cell", "row"),
  fontWeight = NULL,
  color = NULL,
  backgroundColor = NULL,
  background = NULL,
  ...
)
}
\description{
\code{\link[DT]{formatStyle}}
}
