% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{org}
\alias{org}
\title{Example data for TT package}
\format{
data.tree object. Each node has 7 attributes so resulting table will be:\cr
atribute name and months of the year (\code{Jan-2021, Feb-2021....})\cr
each attribute value is some number, parents value is cumulative sum of children
}
\usage{
org
}
\description{
data.tree object that can be used as an example to see how table widget looks like.
Each node of data.tree store attributes (dates) with some numbers.
}
\keyword{datasets}
