% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TTformat.r
\name{formatCurrency}
\alias{formatCurrency}
\title{Format table columns}
\usage{
formatCurrency(
  table,
  columns,
  currency = "$",
  interval = 3,
  mark = ",",
  digits = 2,
  dec.mark = getOption("OutDec"),
  before = TRUE
)
}
\description{
\code{\link[DT]{formatCurrency}}
}
